Fix #2 - Message types should be auto-registered
This change completes the work to have all Transport.Stdio message types
be marked with a MessageTypeNameAttribute.  This attribute specifies
the name used to identify the type in the JSON messages that are sent
between client and editor service.

The MessageTypeResolver also validates whether a message type is decorated
with an attribute and throws a debug assert if it is missing.  This helps
message type authors remember to add the attribute so that everything works
correctly.

Finally, all existing message types have been updated to follow the new
message type decoration pattern.