Eliminate API leakage of some PowerShell types
This change introduces new types that wrap/replace the types
from System.Management.Automation which should not be exposed
outside of the core Editor Services library.  The reasoning is
that a consumer of our higher-level service types through an
EditorSession should not have to take a dependency on the
System.Management.Automation assembly (since this may become
more complicated in the future).  This also gives us a chance to
reshape the information that we return from API calls so that we
can provide extra value in some places and remove unneeded
information in others.

Going forward we will carefully monitor any new function
signatures to ensure that core PowerShell types aren't being
exposed through the abstraction unnecessarily.