Add initial language service tests
This change adds new tests for the LanguageService class.  It also
establishes a new shared library for test assemblies so that both
the core library tests and the host process tests can use the same
test data.  Usage of the shared test data in the host tests will
be added in a later commit.