Add .nuspec files for all shippable projects
This change introduces .nuspec files so that NuGet packages can be
built for all of the shippable projects in the solution.  Initial
package build will be performed by AppVeyor and final package
build for release will be done through some other means after the
binaries have been signed and strong-named.