Add debugging support; refactor session management
This change introduces debugging support to both the core .NET library and
the out-of-process API host.  It adds a new DebugService class to the API
and all of the necessary messages that are specified by VS Code's
debugging protocol.

This change also marks a refactoring of runspace management logic in
the core API.  Previously the runspace used by a few different services
was interacted with directly, which complicated the process for executing
commands and scripts when the debugger was active.  A new
PowerShellSession class has been introduced to provide complete management
of the runspace(s) needed for a single editing session.  In the future,
this class will expand to support remote sessions as well.