Improve ScriptFile loading of trailing newlines
This change fixes a bug in ScriptFile where there would not be a final
empty line if the last content line ended with a newline.  This was
causing some problems when editing the last line of the file in an editor
because the line numbers would be out of range.