Fix breakpoint removal; add "close" request
The first change fixes breakpoint removal.  Previously, breakpoints were
being tracked per-file based on the script's file path.  It was discovered
that file paths were being sent through the protocol with inconsistent
casing, so path normalization was needed to ensure that breakpoints were
being indexed with the correct path casing.

The second change adds the "close" request for closing opened files.  This
request was not being used by VS Code previously but was suddenly getting
sent in newer builds of that editor.