Remove most subnamespaces from core library code
This change removes many subnamespaces that had been used in the core .NET
library for this project.  This should help with type discoverability when
a developer is using IntelliSense.  The only remaining subnamespace is
"Utility" since those classes are not essential to typical usage of the
library.

This change may be reversed in the future if sufficient feedback is
received that subnamespaces are helpful.