Introduce OutputEvent and command eval in debugger
This change introduces the new OutputEvent which allows the the language
and debugging services to send script output text to the client to be
displayed to the user.  Currently this is used in VS Code's Debug Console
but could be used in other contexts as well.

This change also enables the user to evaluate PowerShell commands while
stopped in the debugger to aid in debugging.  In the future this will be
expanded to allow commands to be executed at any time.