Fix #29: Variables should use proper '$' notation
This change adds the dollar sign '$' character to top-level variables that
are retrieved using the Get-Variables command via the DebugService.  This
character is used to refer to variables in PowerShell scripts so we need
to add it to variable names that are returned from the debugging service
for the sake of consistency.