Temporary fix for high CPU usage when stream lost
This change introduces a "fix" for the issue where the language service
host process will spin at high CPU if the client's input stream terminates
unexpectedly.  This issue is caused by incorrectly handling the result of
Stream.ReadAsync which states that a result of 0 means that the stream has
terminated.  Instead of returning true and continuing the read loop, I now
throw an EndOfStreamException which causes the app to terminate with an
unhandled exception.

A future fix will be implemented to provide a more graceful shutdown path.