Catch exception in VariableDetails.GetChildren()
This change catches the GetValueInvocationException that can be raised in
VariableDetails.GetChildren() when it is called on a thread other than the
debugger pipeline thread.  In the future, all requests for children of a
variable should be done on this thread.  For now we just catch the
exception so that the debug adapter does not crash.