Fix #46: Clear-Host causes error while debugging
This change updates the SessionPSHostRawUserInterface class to get rid of
most instances of throwing NotImplementedException.  When debugging, if a
script invokes a method that causes the raw UI interface to be touched,
those methods shouldn't cause errors to appear.  This change reduces the
possibility of the user seeing errors due to this reason.