Add new client/server API for host process
This change introduces a new set of classes that make it much easier to
write a language server or debug adapter and clients for either type of
server.  All client/server logic has been moved to the Protocol library so
that it can be used in other host applications outside of the provided
host process.