Add AppVeyor RDP config to appveyor.yml
This change adds a couple lines to appveyor.yml that can be used to enable
RDP access to the build server when something goes wrong.  To use this
configuration, the developer must uncomment the two lines and make another
commit to their PR branch.  Details are included in the config file.