Fix hang when debug adapter finishes executing
This change fixes a hang that was caused by the recent refactoring of host
process server implementations.  In the debug adapter implementation, the
PowerShellContext was not being disposed at shutdown which was causing the
process to stay open because the associated threads were still alive.
This change adds the necessary ProtocolServer.Shutdown implementation.