Fix #47: Add 'context' param to EvaluateRequest
This change adds the next 'context' parameter to EvaluateRequest, allowing
the debug adapter client to identify whether an 'evaluate' request will be
used for evaluating an expression for value display or REPL execution.
This allows us to write out REPL command output only when expressions are
being evaluated in that context.