Stream output from commands run in debugger
This change updates the way we execute commands in the debugger which need
to have their output streamed to the console.  Previously the command
execution would just evaluate the expression and return the results
without writing them to the host.  Now for PowerShell v4/5 users, the output
can be streamed as the command is being executed.  For PowerShell v3 users
the output will be written after the command fully completes executing.