Fix falsely passing PowerShellVersionTests
This change fixes the current PowerShellVersionTests to use an absolute
path to the version-specific PowerShell reference assembly.  The relative
path which was being used somehow caused MSBuild to use the in-box
PowerShell v5 assemblies since the HintPath couldn't be found.  This
caused the tests for PowerShell 3 and 4 to pass even though they should be
failing.

This change also includes some fixes to existing code that is incompatible
with PowerShell 3 and 4.