Add choice prompt handling for PSHostUserInterface
This change adds a general solution for handling the PromptForChoice
method in our PSHostUserInterface implementation.  It abstracts the
display and response handling for prompts so that different prompt
interfaces can be used in different contexts, all within the same session.

The initial implementation includes a console-based prompt handler which
operates much like the standard prompt display in the PowerShell console.
Another implementation is provided which sends an event through the JSON
protocol so that an editor client can display the prompt in an
appropriate user interface.

Resolves #85.