Send EvaluateRequest after execution completes
This change modifies the way the EvaluateRequest message is handled in the
LanguageServer.  Previously the evaluation response would be sent back
immediately since we aren't giving back an actual result.  This behavior
is now changed to wait until execution completes before sending the
response so that the caller knows when execution has completed.