Fix #130: Reduce prompt noise in debug adapter
This change keeps the debug adapter from printing the prompt until the
user enters a command.  This will allow breakpoints to be hit without the
prompt being shown every time.  The change also trims the ending '>'
character off of the end of the prompt to reduce user confusion about
where commands should be typed.