Add intervallic buffering of OutputEvent messages
This change introduces a new AsyncDebouncer class which is able to gate
invocation requests to only occur after a specified interval.  This class
is used to implement an OutputDebouncer which takes all output messages
written from the PowerShell host and buffers them.  This causes many lines
of output written in a short time window to be sent as one OutputEvent
rather than writing individual events for each line.

Resolves #113.