Fix #147: Enable native console app output in REPL
This change enables native console app output to be written in the host
process without it breaking the standard I/O channel between the client
and host.  It accomplishes this by setting an internal property in the
PowerShell runtime called ConsoleVisibility.AlwaysCaptureApplicationIO.
This is a short term solution until we find a better way to deal with
native application output.