More work on #149.  Need to be able to unescape paths as well as escape the.
The trick is to unescape paths coming from VSCode as the user may have decided to escape spaces and [ and ].  .NET APIs won't handle those paths.  So we unescape them and cache them in Workspace.cs.  Later, just before we call down to the PowerShell engine to do work that requires the script path, we escape those paths appropriately for PowerShell.