Fix #159: LanguageServer.Shutdown hangs on flush
The change fixes an issue in the LanguageServer's Shutdown method where
buffered output in the OutputDebouncer gets flushed before shutting down.
The problem here is the Wait call which blocks the message dispatcher
thread from completing the OutputDebouncer.OnFlush method's SendEvent
call.  The fix is to change the Shutdown method to async so that the
Flush call can be awaited, unblocking the message dispatcher thread's
SynchronizationContext for the SendEvent call to complete.