Fix #153: Script Analyzer is not returning markers
This change fixes an issue in 0.4.1 where Script Analyzer analysis markers
are not being returned by the AnalysisService.  This issue was caused by a
new optional parameter to the ScriptAnalyzer.Initialize method called
'includeDefaultRules' which was introduced in 1.3.0.  Since this parameter
has a default value of 'false', the default Script Analyzer rules were not
being loaded or applied in analysis runs.