Fixes #173 - implementation for function breakpoints
This commit also introduces support for the configurationDone request.  In fact, due to the order in which set*Breakpoints requests and the launch request come in, we now defer the execution of the script until we receive the configurationDone request.  When we executed from the launch request, the debug host would appear to hang.  It did not like executing Set/Remove-PSBreakpoint commands when the host debugger was running (not in a stopped state) AFAICT.

This commit has a known issue with VSCode mixing up which breakpoint has a condition associated with it when you add/remove a breakpoint (see https://github.com/Microsoft/vscode/issues/3558).  The next commit will fix this issue.