This is a *proposed* change to address the issue that the launch request could come after the configurationDone request (https://github.com/Microsoft/vscode/issues/3548).
This commit "should" allow the messages to come in either order and still work.  However, it needs a good review.  BTW it appears all the messages are processed on the same thread (or maybe that was coincidence).  If that is the case then perhaps the lock is not needed.  Still it is a fairly lightweight lock - one bool check and an assignment.