Provide a bit more accurate "reason" for debug stop.
It turns out that when the debugger stops at a function breakpoint, it isn't necessarily obvious why the debugger stopped.  There is no breakpoint glyph for a function breakpoint.  So if we return "function breakpoint", then VSCode will display "PAUSED ON FUNCTION BREAKPOINT" in the call stack viewlet.  Hopefully that will help the user know why the debugger stopped.