Fix #181: Fix crash for unsupported field types
This change adds proper error handling for input prompt requests for field
types that we don't yet support like SecureString and PSCredential.
Previously any Read-Host or Get-Credential call which prompts for either
of these types would crash the language and debugging services.  The fix
causes errors to be written to the host stating that these commands and
field types are not yet supported.