Remove unneeded newlines when stepping in debugger
This change removes a newline that was being written to the debug console
when the user steps through code in the debugger.  This newline was added
for the case where the debug prompt is written at each stop but was not
removed when this feature was temporarily taken out.  The newline will be
reinstated once we have an option for turning on the debugger prompt.