Introduce new editor extensibility API
This change introduces a new PowerShell-based API that allows scripts and
modules to extend the behavior of the host editor.  Scripts can either
automate the editor directly or register commands that can later be
invoked by the user.  The ExtensionService provides this behavior for any
PowerShellContext by injecting a new variable '$psEditor' into the
session.

This version of the API is very early and will be expanded upon in future
releases.