Fix broken profile loading test
This change fixes the LanguageServerTests.ServiceLoadsProfilesOnDemand
test which was broken by recent changes to our setting loading code.  It
removes a potential condition where sending a null string for the Script
Analyzer settings path causes the AnalysisService to be restarted.  This
restart resulted in a load failure which polluted the console output,
causing the expectations for the profile loading test to fail.