Add generated documentation site
This change introduces a new documentation site for the PowerShell Editor
Services project.  It uses the DocFX tool to generate both conceptual and
API documentation pages.  The conceptual documentation pages come from
Markdown files in the docs folder.  The API documentation pages are
generated dynamically from the XML documentations in the C# classes of the
main .NET DLL of the project.

A script called BuildDocs.ps1 is included to handle the generation and
publishing of the documentation site.