Convert Editor Services host to PowerShell module
This (unfortunately large) commit includes two important design changes to
the PowerShell Editor Services hosting model.

The first change is the conversion of the host process to be loaded as a
PowerShell module inside of powershell.exe.  This change enables a lot of
interesting possibilities including the ability to run the Editor Services
host on a remote machine and connect to it from a client machine.  Another
benefit of this approach is the ability to load assembly dependencies
based on the version of PowerShell in which the host is loaded.

Following on from the previous point, the second change is to introduce a
separate build of the PSES host which is compiled against the Nano Server
SDK.  This enables the PSES host to be loaded in the edition of PowerShell
that is shipped with Nano Server.

Both of these changes bring us much closer to enabling a remote editing
experience which would allow someone to use VS Code on a Linux or OSX
machine and connect to an Editor Services host running in a Windows or
Nano Server VM.  We will need to finish the WebSockets channel support
before this is fully possible, though.