Fix #221: Occasional init hang in LanguageServer
This change fixes an issue in the LanguageServer where the length of time
that it takes to evaluate the extension API script can cause the
"initialize" request to not complete.  This leaves the VS Code language
server client in an initializing state where no language features will
work for the rest of the session.  The fix is to execute the extension API
script after the "initialize" handler has been registered.