Fix #222: Prompt functions hang in editor commands
This change fixes an issue where $host.UI prompt functions do not return
the user's response when run inside of PowerShell editor commands.  This
was caused by the HandleInvokeExtensionCommand method returning a Task
that did not complete until after the command completed.  The fix was to
not await on completion of the command so that the prompt response has a
chance to come through.