Need to escape regex pattern with external component.
LanguageServer.CreateCompletionItem was causing a editor host crash because it (er *I*) was injecting tooltip text (returned by the PS engine) into a regex pattern without escaping it.  One of the completion items was 'Notepad++.exe' which resulted in a regex parsing error.  Would like to add a test for this at some point but need to think on it.  CreateCompletionItem is private.

Fixes #193