Provide temporary separate .sln for CI builds
This change introduces a separate PowerShellEditorServices.NoNano.sln
solution file so that AppVeyor will not build the Nano Server version of
the project.  This is a temporary measure until we get a build image with
the Nano Server SDK baked in.  After that we will take this solution file
back out so that the Nano Server version gets built by AppVeyor.