Fix #264: Variable and parameter IntelliSense broken in VS Code 1.4.0
In VS Code 1.4.0, the completion behavior has changed to require the use
of the filterText field to ensure that completions which insert text not
equal to the label field will get displayed. Since PowerShell often
suggests completions with labels different than the actual insert string,
this caused some completions to be broken.

The fix is to put the value of the insertText field into the filterText
field also.