Fix #284: AbortException crashes when called more than once
This change fixes an issue in PowerShellContext.AbortException which
causes it to throw a NullReferenceException when called after the session
has already been aborted.  This issue manifests sometimes when the
debug adapter finishes execution normally or when it terminates due to a
script parsing error.  It was initially reported in the following two
issues for the PowerShell VS Code extension:

PowerShell/vscode-powershell#243
PowerShell/vscode-powershell#273

The fix is to check whether the PowerShellContext.SessionState is Aborted
or Disposed before proceeding with the abort.