Fix issue with debugger output not being written for short scripts
This change fixes an issue where script output for very small scripts is
not being written out reliably before the debug adapter terminates.  The
fix is to add an extra output flush before sending the TerminatedEvent
back to the client.  This fix is temporary until we get the REPL
integration online.

Resolves #138.
Resolves PowerShell/vscode-powershell#284.