Fix #325: Module dependency loading should pick latest module version
This change fixes an issue with the loading of both the PSScriptAnalyzer
and Plaster modules.  We were simply using Import-Module without
specifying a version to be loaded.  Apparently PowerShell prioritizes the
PSModulePath order over the module version order so there are instances
where an older version of one of these module will be loaded even though a
newer version is available at a different location on the machine.

This change adds extra logic to sort the list of returned modules so that
the latest version of the module is always loaded.