Enable debugging of remote and attached runspaces
This change finishes support for Enter-PSSession and Enter-PSHostProcess
by wiring up the AttachRequest so that both of these commands can be used
to attach to a runspace in a process on a local or remote machine.  It
also enables the LaunchRequest to be executed without a script file path
to enable an interactive debugging session experience.