Handle PSSA settings hashtable input
This enables the client to send settings as a hashtable instead of pointing to a settings file for script analysis. Our current code structure allows us to pass either a set of rules or a settings file path but not both. However, to provide code formatting settings we not only need to send a settings file to configure the given rules but also the rules in a particular order. The best approach in this case is to update and send a settings hashtable everytime we request code formatting markers.