Add RemoteFileManager to manage remote files
This change adds a new RemoteFileManager class which manages files which
are opened when a remote session is active.  It handles the gathering of
remote file contents and mapping remote files paths to a local cache.  It
also keeps track of the remote files which were opened in the editor and
closes them once the remote session has been exited.