Implement psedit command for remote file loading
This change adds new behavior to RemoteFileManager so that a 'psedit'
command will be added to all sessions for the purpose of loading local or
remote files in the debugging experience.  For now the remotely opened
files do not have new contents propagated to the remote machine but that
will be added shortly.