Add !ctrlc and !break commands to debug adapter evaluator
This change adds the !ctrlc and !break commands which are special to the
debug adapter's REPL interface.  The provide the ability to send a Ctrl+C
(abort) or Ctrl+B (break) at any time.  This is necessary because the
debugger UI can't properly handle or transfer these commands in some
cases.