Handle non-string return values from the user's prompt function
This change adds a catch for the InvalidCastException when gathering the
user's prompt string in the case where their prompt function returns a
value other than a string.  In this case a default prompt is returned.

This change also improves the way we write prompts for the debugger.
Instead of writing out the prompt each time the debugger stops, we write
out the details about the breakpoint that was hit, if any.  We also try to
preserve any session context at the beginning of the prompt string like
the remote computer name or process ID.

Fixes #339.