Fix #346: GetVersionDetails failing due to empty environment variable
This change fixes an issue caused by a recent change in the VS Code
extension which wiped out all of the environment variables in the process.
That particular issue has now been fixed but it's still good to guard
against this problem in the future.  This change adds a null check to
ensure that we don't throw a NullReferenceException in future occurrences
of the issue.