Fix PSScriptAnalyzer module loading in new async AnalysisService
This change fixes the way that the PSScriptAnalyzer module gets loaded in
the new asynchronous AnalysisService.  Since we changed to using the
InitialSessionState to load the module, we were unable to specify the
exact version of the module we needed.  This change goes back to an
explicit Import-Module across the runspaces in the RunspacePool.