Update to SDK version 1.0.0-rc3-004517
This change updates the .build.ps1 script to point at the 1.0.0-rc3
release of the .NET SDK so that we can make use of updated test command
that returns the correct exit code when tests fail.  The build script is
also updated to pull the SDK version from global.json.  This file is also
used to tell the dotnet CLI tool which SDK version to use.