Use the right SendEvent call in server implementations
This change fixes broken language and debug server tests by switching to a
different SendEvent implementation which ensures that the event is being
sent from the correct thread.  A true fix will be needed in the future to
ensure that all SendEvent implementations will use the correct thread for
sending event messages.