Use only 1 runspace in analysis service
Having multiple runspaces crashed scriptanalyzer invocation because of the following two issues:
* PSCmdlet.WriteOutput being called from another thread than Begin/Process
* CompositionContainer.ComposeParts complaining that "...Only one batch can be composed at a time"

We used multiple runspaces in the first pace because PSSA was slow and having multiple instances sped things up. But having multiple runspaces is causing a lot of issues which are hard to debug. Also, PSSA has become significantly fast compared to the previous version and there having just one runspace should give a reasonable performance. And on the positive side we do not have to deal with the aforementioned crashes.