Improve detection of runspace changes
This change adds new behavior that makes the detection of sudden runspace
changes much more reliable.  These changes were particularly necessary for
detecting when we start debugging a runspace in a remote session since we
don't get a PushRunspace notifications when that happens.

This change also improves how we gather the prompt string just after a
command is executed.