Fix #357: Watch evaluation causes debugger hang
This issue is caused when VS Cod sends watch expression evaluation
requests after the debugger has been resumed.  This issue has been fixed
in the most recent builds of VS Code but we must have a guard for this on
our side for those who are still using VS Code Stable.