Modify build script's Build task to not build solution
This change modifies the build script so that it doesn't build the entire
solution file.  This is needed because the test projects depend on net452
now due to xUnit requirements; building the solution with both net451 and
net452 was causing compilation errors, so it was better to just take that
out and let "dotnet test" build the test projects on demand.