Enable saving remote files opened with psedit
This change enables the saving of files opened from a remote session using
the psedit command.  When the editor saves the file contents it sends a
DidSaveTextDocumentNotification to the language server which in turn sends
the document's contents back to the remote session to be saved in the
original file.

Resolves PowerShell/vscode-powershell#583.