Interrupt integrated console command prompt when commands are executed
This change fixes a general class of problems in the integrated console
experience where commands that were executed internally (like extension
commands or running Plaster templates) do not interrupt the command
prompt, causing future input to be written over previous output lines.

The fix is to have the command prompt be interrupted by any commands
which write output or errors to the host.

Fixes #411.