Fix debugger startup hang when session still initializing
This change fixes a hang that occurs when the user starts the debugger
right when the PowerShell session is still initializing.  This is caused
by a call to PowerShellContext.SetWorkingDirectory which isn't
respecting the current state of the runspace.  The fix is to ask for a
RunspaceHandle to ensure that the runspace is free before attempting to
set the working directory.

Resolves PowerShell/vscode-powershell#632