Enable debugging of PowerShell Editor Services using VS Code
This change turns on the portable PDB format for the primary assemblies
in the project so that they can be debugged in VS Code.  Version 1.9 of
the C# extension added a new debugger type, `clr`, which allows you to
debug 64-bit Full CLR apps when portable PDBs are in use.