Remove while loop used for reading error stream
Whenever EditorServices fails to start up, ServerTestsBase.LaunchService goes into a while loop where it reads from the error stream line by line in an async manner, till the end of the stream. This causes the method to wait indefinitely in certain situations. Hence, we replace it with ReadToEndAsync method and time out on the task if it takes too long. This avoids the test cases from hanging at that point.