Fix command prompt cursor location issues while debugging
This change fixes an issue which causes the user's cursor location to
jump around when they type the first letter after stepping through code
or hitting multiple breakpoints sequentially.  This is caused by a
hanging input listener which wasn't respecting the cancellation of its
task when it gets evaluated.

Fixes PowerShell/vscode-powershell#698.