Fix debugger hang when you stop execution while input prompt is active
This change fixes an issue where the debugger will hang when the user
tries to stop execution while there is an active input or choice prompt.
The solution is to use a more robust method for cancelling the prompt
task which doesn't depend on the Console.ReadKey() call to return.

Fixes #428.