Breakpoints hit when debugger isn't active now notify editor
This change is part of the fix for PowerShell/vscode-powershell#619
which states that hitting a breakpoint in the integrated console does
not activate the debugger in VS Code.  The fix is to check whether a
debugger client is connected when a breakpoint is hit, and if not, send
a notification through the language server to have the editor connect
its debugger client.