Move runspace creation outside of PowerShellContext
This change moves runspace creation outside of PowerShellContext so that
the runspace can be configured more easily at the host level.  It also
changes the relationship between the ConsoleServicePSHost and
ConsoleService so that the latter now gets created by the former.  This
is leading up to a change where the behavior of these two classes will
be merged into the PSHost implementation.