Write custom full prompt each time it is displayed
This change fixes an issue where the user's custom prompt function is
only written out fully (including any Write-Host or Write-Output
behavior) the first time the prompt appears after a command is executed.
This is because the return value of the prompt function was being cached
and re-printed each time the user hit Enter without running a new
command.

Resolves PowerShell/vscode-powershell#774.