Improve reliability of execution on pipeline thread while debugging
This change improves reliability of command execution on the pipeline
thread when stopped at a breakpoint in the debugger.  Previously there
was a global task variable used to propagate execution results back to
the caller of ExecuteCommand, but now we use a specific task variable
that's tied to the original ExecuteCommand call to avoid reuse of the
global task.