Fix evaluation of prompt function output to remove unneeded "PS>"
This change fixes an issue in the integrated console where the result of
a user's custom prompt function would not be interpreted correctly,
causing the default prompt string of "PS>" to be written out when it
wasn't needed.  The fix is to expect a PSObject result rather than
object so that we can evaluate the inner object's type correctly.

Resolves PowerShell/vscode-powershell#808.