Decouple ProtocolEndpoint from client and server implementations
This change decouples the ProtocolEndpoint class from the existing
client and server implementations for the language and debug services.
The goal here is to eventually decentralize all editor feature
implementations into individual classes, so separating out the
ProtocolEndpoint is the first step in achieving this.  This class will
now be exposed in the framework as the IMessageSender interface.

This change also simplifies the LanguageServer/DebugAdapter client and
server implementations, removing some unnecessary abstraction and base
classes.