Add position classes for new functions
This change adds a custom implementation of IScriptExtent similar to
InternalScriptExtent.  The major difference being the replacement of
PositionHelper with FileContext.  The PositionHelper class was used to
store the full script text and line start mapping for all ScriptExtent
objects of the same file.  FileContext is a great replacement for this
because it's already kept unique by PSES and it adapts to changes.

- Added FullScriptExtent and FullScriptPosition classes.

- Changed the scriptFile field on FileContext from private to internal.